feat: demo attribute changes
This shows how you would have the React app in the web component update
in response to attribute changes. As you can see, requires a bit of
manual wiring up. I turned the `Greeter` component into a controlled
component so that I could set the name from outside or via the input.